Add getInfo GetMappings to Author, User; delegate Repositories' work to Services

+++ src/main/java/com/onlib/core/service/ReviewService.java
@@ -1,5 +1,6 @@
 package com.onlib.core.service;
 
+import com.onlib.core.dto.ReviewWithoutBookDto;
 import com.onlib.core.model.Book;
 import com.onlib.core.model.Review;
 import com.onlib.core.model.LibraryUser;
@@ -8,9 +9,10 @@ import com.onlib.core.repository.ReviewRepository;
 import com.onlib.core.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
+import java.util.List;
 
 @Service
 public class ReviewService {
@@ -25,15 +27,15 @@ public class ReviewService {
 
     /// Добавляется ли автоматически review в book?
     @Transactional
-    public void addReview(String text, Long userId, Long bookId) {
+    public void addReview(String text, Long userId, Long bookId) throws NotFoundException {
         Book book = bookRepository.findById(bookId).orElse(null);
         if (book == null) {
-            throw new RuntimeException("Can't find book with id " + bookId);
+            throw new NotFoundException();
         }
 
         LibraryUser libraryUser = userRepository.findById(userId).orElse(null);
         if (libraryUser == null) {
-            throw new RuntimeException("Can't find user with id " + userId);
+            throw new NotFoundException();
         }
 
         Review review = reviewRepository.save(new Review(text));
@@ -43,7 +45,18 @@ public class ReviewService {
     }
 
     @Transactional
-    public void deleteReview(Long id) {
-        reviewRepository.deleteById(id);
+    public void deleteReview(Long id) throws NotFoundException {
+        if (reviewRepository.existsById(id))
+            reviewRepository.deleteById(id);
+        else
+            throw new NotFoundException();
+    }
+
+    @Transactional
+    public List<ReviewWithoutBookDto> getReviewsWithoutBookByBookId(Long bookId) {
+        List<Review> reviews = reviewRepository.findByBookId(bookId);
+        return reviews.stream()
+                .map(ReviewWithoutBookDto::new)
+                .toList();
     }
 }
